modules/chucklesweb/requirements.txt
azure-iothub-device-client~=1.4.3
Flask==0.10.1
RPi.GPIO>=0.6.3